+++ src/main/java/com/clubber/ClubberServer/domain/admin/service/AdminEmailAuthService.java
@@ -7,7 +7,7 @@ import com.clubber.ClubberServer.domain.admin.dto.UpdateAdminUpdateEmailAuthVeri
 import com.clubber.ClubberServer.domain.admin.exception.AdminInvalidAuthCodeException;
 import com.clubber.ClubberServer.domain.admin.implement.AdminReader;
 import com.clubber.ClubberServer.domain.admin.repository.*;
-import com.clubber.ClubberServer.domain.admin.validator.AdminValidator;
+import com.clubber.ClubberServer.domain.admin.implement.AdminValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
